api/pyproject.toml
[tool.poetry]
name = "parcelec"
version = "0.1.0"
description = ""
authors = ["thomas-god <thomas.godlewski@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pendulum = "^2.1.2"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.9.5"
isort = "^5.9.2"
black = "^21.7b0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
